# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Financial Coach API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analysis/run:
    post:
      operationId: runAnalysis
      tags: [analysis]
      summary: Run multi-agent financial analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialInput"
      responses:
        "200":
          description: Analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Analysis failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /analysis/history/{id}:
    get:
      operationId: getAnalysis
      tags: [analysis]
      summary: Get a specific past analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteAnalysis
      tags: [analysis]
      summary: Delete a past analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/signup:
    post:
      operationId: signupUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Log in an existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: User logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Verify the active session and get current user details
      responses:
        "200":
          description: Active session user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Session invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /billing/create-subscription:
    post:
      operationId: createSubscription
      tags: [billing]
      summary: Create a new Razorpay subscription
      responses:
        "200":
          description: Subscription created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptionId:
                    type: string
                  keyId:
                    type: string
                required: [subscriptionId, keyId]
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /billing/verify-signature:
    post:
      operationId: verifySignature
      tags: [billing]
      summary: Verify the Razorpay payment signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                razorpayPaymentId:
                  type: string
                razorpaySubscriptionId:
                  type: string
                razorpaySignature:
                  type: string
              required: [razorpayPaymentId, razorpaySubscriptionId, razorpaySignature]
      responses:
        "200":
          description: Signature verified and user upgraded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    DebtItem:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        interestRate:
          type: number
        minPayment:
          type: ["number", "null"]
      required: [name, amount, interestRate]
    FinancialInput:
      type: object
      properties:
        monthlyIncome:
          type: number
        dependants:
          type: integer
        expenses:
          type: object
          additionalProperties:
            type: number
        debts:
          type: array
          items:
            $ref: "#/components/schemas/DebtItem"
        notes:
          type: ["string", "null"]
      required: [monthlyIncome, expenses]
    SpendingCategory:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        percentage:
          type: number
      required: [category, amount, percentage]
    SpendingRecommendation:
      type: object
      properties:
        category:
          type: string
        recommendation:
          type: string
        potentialSavings:
          type: ["number", "null"]
      required: [category, recommendation]
    BudgetAnalysis:
      type: object
      properties:
        totalExpenses:
          type: number
        monthlyIncome:
          type: ["number", "null"]
        savingsRate:
          type: ["number", "null"]
        spendingCategories:
          type: array
          items:
            $ref: "#/components/schemas/SpendingCategory"
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/SpendingRecommendation"
        summary:
          type: string
      required: [totalExpenses, spendingCategories, recommendations, summary]
    EmergencyFund:
      type: object
      properties:
        recommendedAmount:
          type: number
        currentStatus:
          type: string
        monthsOfExpenses:
          type: number
      required: [recommendedAmount, currentStatus, monthsOfExpenses]
    SavingsRecommendation:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        rationale:
          type: string
      required: [category, amount, rationale]
    SavingsStrategy:
      type: object
      properties:
        emergencyFund:
          $ref: "#/components/schemas/EmergencyFund"
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/SavingsRecommendation"
        automationTips:
          type: array
          items:
            type: string
        summary:
          type: string
      required: [emergencyFund, recommendations, automationTips, summary]
    PayoffMethod:
      type: object
      properties:
        name:
          type: string
        totalInterest:
          type: number
        monthsToPayoff:
          type: integer
        monthlyPayment:
          type: ["number", "null"]
        order:
          type: array
          items:
            type: string
      required: [name, totalInterest, monthsToPayoff, order]
    DebtReduction:
      type: object
      properties:
        totalDebt:
          type: number
        avalanche:
          $ref: "#/components/schemas/PayoffMethod"
        snowball:
          $ref: "#/components/schemas/PayoffMethod"
        recommendations:
          type: array
          items:
            type: string
        summary:
          type: string
      required: [totalDebt, recommendations, summary]
    AnalysisResult:
      type: object
      properties:
        id:
          type: ["integer", "null"]
        createdAt:
          type: ["string", "null"]
        input:
          $ref: "#/components/schemas/FinancialInput"
        budgetAnalysis:
          $ref: "#/components/schemas/BudgetAnalysis"
        savingsStrategy:
          $ref: "#/components/schemas/SavingsStrategy"
        debtReduction:
          $ref: "#/components/schemas/DebtReduction"
      required: [input, budgetAnalysis, savingsStrategy, debtReduction]
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        employmentStatus:
          type: ["string", "null"]
        annualIncomeRange:
          type: ["string", "null"]
        riskTolerance:
          type: ["string", "null"]
        baseCurrency:
          type: ["string", "null"]
        subscriptionStatus:
          type: ["string", "null"]
        subscriptionTier:
          type: ["string", "null"]
        currentPeriodEnd:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        trialDaysRemaining:
          type: integer
        isTrialExpired:
          type: boolean
      required: [id, email, firstName, lastName, createdAt, updatedAt, trialDaysRemaining, isTrialExpired]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    SignupInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        employmentStatus:
          type: ["string", "null"]
        annualIncomeRange:
          type: ["string", "null"]
        riskTolerance:
          type: ["string", "null"]
        baseCurrency:
          type: ["string", "null"]
      required: [email, password, firstName, lastName]
...